Fix NaN in Spark functions
