+++ velox/functions/sparksql/ArrayIntersectExcept.cpp
@@ -1,0 +1,61 @@
+/*
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+#include "velox/functions/lib/ArrayIntersectExceptFunction.h"
+
+namespace facebook::velox::functions::sparksql {
+namespace {
+
+std::shared_ptr<exec::VectorFunction> createArrayIntersect(
+    const std::string& name,
+    const std::vector<exec::VectorFunctionArg>& inputArgs,
+    const core::QueryConfig& /*config*/) {
+  validateMatchingArrayTypes(inputArgs, name, 2);
+  auto elementType = inputArgs.front().type->childAt(0);
+
+  return VELOX_DYNAMIC_SCALAR_TEMPLATE_TYPE_DISPATCH(
+      createTypedArraysIntersectExcept,
+      /* isIntersect */ true,
+      elementType->kind(),
+      inputArgs,
+      true);
+}
+
+std::shared_ptr<exec::VectorFunction> createArrayExcept(
+    const std::string& name,
+    const std::vector<exec::VectorFunctionArg>& inputArgs,
+    const core::QueryConfig& /*config*/) {
+  validateMatchingArrayTypes(inputArgs, name, 2);
+  auto elementType = inputArgs.front().type->childAt(0);
+
+  return VELOX_DYNAMIC_SCALAR_TEMPLATE_TYPE_DISPATCH(
+      createTypedArraysIntersectExcept,
+      /* isIntersect */ false,
+      elementType->kind(),
+      inputArgs,
+      true);
+}
+} // namespace
+
+VELOX_DECLARE_STATEFUL_VECTOR_FUNCTION(
+    udf_array_intersect,
+    signatures("array({})"),
+    createArrayIntersect);
+
+VELOX_DECLARE_STATEFUL_VECTOR_FUNCTION(
+    udf_array_except,
+    signatures("array({})"),
+    createArrayExcept);
+} // namespace facebook::velox::functions::sparksql
